refactor: better common/js
