+++ js/common/array.ts
@@ -1,5 +1,7 @@
-// 拖拽排序场景中：调整某个元素的顺序
-export default function swapDragArrayElement(data: any[], currentIndex: number, targetIndex: number) {
+/**
+ * 拖拽排序场景中：调整某个元素的顺序
+ */
+export function swapDragArrayElement(data: any[], currentIndex: number, targetIndex: number) {
   const newData = [...data];
   if (targetIndex - currentIndex > 0) {
     newData.splice(targetIndex + 1, 0, newData[currentIndex]);
