+++ js/common/index.ts
@@ -1,0 +1,32 @@
+export * from './types';
+export * from './constants';
+export * from './components';
+export {
+  observe,
+  setStyle,
+  injectStyle,
+  getScrollbarWidth,
+  getScrollbarWidthWithCSS,
+  getIEVersion,
+  getFlexGapPolyFill,
+  calculateNodeSize,
+  calcTextareaHeight,
+  getColorTokenColor,
+} from './dom';
+export { linear, easeInOutCubic } from './number';
+export type { EasingFunction } from './number';
+export { template, firstUpperCase, getCharacterLength, getUnicodeLength, limitUnicodeMaxLength } from './string';
+export {
+  hasOwn,
+  getPropertyValFromObj,
+  isPlainObject,
+  isPromise,
+  omit,
+  getValidAttrs,
+  removeEmptyAttrs,
+  getTabElementByValue,
+} from './object';
+export { swapDragArrayElement } from './array';
+export { getBackgroundColor, pxCompat } from './style';
+export type { Gradients, FromTo, LinearGradient } from './style';
+export { getPosition, calcSize } from './layout';
