+++ js/global-config/index.ts
@@ -1,0 +1,10 @@
+export * from './t';
+export * from './default-config';
+export * from './locale/zh_CN';
+export * from './locale/en_US';
+export * from './locale/ja_JP';
+export * from './locale/ko_KR';
+export * from './locale/ru_RU';
+export * from './locale/zh_TW';
+export * from './locale/ar_KW';
+export * from './locale/it_IT';
