+++ js/input-number/index.ts
@@ -1,0 +1,2 @@
+export * from './large-number';
+export * from './number';
