+++ js/loading/index.ts
@@ -1,0 +1,1 @@
+export { default as circleAdapter } from './circle-adapter';
