+++ js/qrcode/index.ts
@@ -1,0 +1,3 @@
+export * from './qrcodegen';
+export * from './types';
+export * from './utils';
