+++ js/statistic/index.ts
@@ -1,0 +1,3 @@
+export { default as Tween } from './tween';
+export type { TweenSettings } from './tween';
+export * from './utils';
