+++ js/statistic/tween.ts
@@ -4,7 +4,7 @@
  * TDesign vue 2 need to ensure compatibility with users who are using IE and Vue2,
  * it is necessary to use setInterval instead of requestAnimationFrame when the browser version is less than 9
  */
-import { getIEVersion } from '../utils/helper';
+import { getIEVersion } from '../common';
 
 export interface TweenSettings {
   from: Record<string, number>;
