+++ js/table/index.ts
@@ -1,0 +1,3 @@
+export { default as TableTreeStore } from './tree-store';
+export * from './types';
+export * from './utils';
