+++ js/tree-v1/index.ts
@@ -1,0 +1,4 @@
+export { default as TreeNode } from './tree-node';
+export { createNodeModel, updateNodeModel } from './tree-node-model';
+export { default as TreeStore } from './tree-store';
+export * from './types';
