+++ js/tree/index.ts
@@ -1,0 +1,4 @@
+export { default as TreeNode } from './tree-node';
+export { TreeNodeModel, nodeKey, createNodeModel } from './tree-node-model';
+export { default as TreeStore } from './tree-store';
+export * from './types';
