+++ js/upload/index.ts
@@ -1,0 +1,4 @@
+export * from './main';
+export * from './types';
+export * from './utils';
+export * from './xhr';
