--- js/utils/injectStyle.ts
@@ -1,9 +1,0 @@
-const injectStyle = (style: string) => {
-  const styleElement = document.createElement('style');
-  let styleSheet = null;
-  document.head.appendChild(styleElement);
-  styleSheet = styleElement.sheet;
-  styleSheet.insertRule(style, styleSheet.cssRules.length);
-};
-
-export default injectStyle;
