+++ js/watermark/index.ts
@@ -1,0 +1,3 @@
+export { default as generateBase64Url } from './generateBase64Url';
+export { default as randomMovingStyle } from './randomMovingStyle';
+export * from './type';
